# apidoc/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "CatCafeApi",
    "description": "猫カフェのポータルサイトと管理画面で使用するAPIを定義",
    "version": "0.1.0"
  },
...
  "servers": [
    {
      "url": "http://{username}:{port}/{basePath}",
      "description": "ローカル環境",
      "variables": {
        "username": {
          "default": "localhost"
        },
        "port": {
          "enum": [
            "8080"
          ],
          "default": "8080"
        },
        "basePath": {
          "default": ""
        }
      }
    }
  ],
  "paths": {
    "/{cats}/notice": {
      "summary": "",
      "description": "",
      "parameters": [
        {
          "name": "cats",
          "in": "path",
          "description": "カフェ識別子",
          "required": true,
          "schema": {
            "type": "string",
            "enum": [
              "cats"
            ],
            "x-enum-varnames": [
              "cats"
            ]
          }
        }
      ],
      "post": {
        "tags": [
          "notice"
        ],
        "summary": "お知らせ登録API",
        "description": "お知らせを1件登録する",
        "operationId": "noticeCreate",
        "requestBody": {
          "description": "リクエストボディ",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "./schemas/notice/request.json#/definitions/notice_create"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "正常系",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./schemas/success_response/postResponse.json#/definitions/post_object"
                }
              }
            }
          },
          "400": {
            "description": "バリデーションエラー",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./schemas/error_response/validationResponse.json#/definitions/validation_result"
                }
              }
            }
          }
        }
      }
    },
    "/{cats}/notices": {
      "summary": "",
      "description": "",
      "parameters": [
        {
          "name": "cats",
          "in": "path",
          "description": "カフェ識別子",
          "required": true,
          "schema": {
            "type": "string",
            "enum": [
              "cats"
            ],
            "x-enum-varnames": [
              "cats"
            ]
          }
        }
      ],
      "get": {
        "tags": [
          "notice"
        ],
        "summary": "お知らせ一覧取得API",
        "description": "お知らせを取得する",
        "operationId": "noticeSearch",
        "parameters": [
          {
            "name": "store_ids",
            "in": "query",
            "description": "店舗ID",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "integer",
                "format": "int32"
              }
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "取得サイズ",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "正常系",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./schemas/notice/response.json#/definitions/notice_search_response"
                }
              }
            }
          },
          "400": {
            "description": "バリデーションエラー",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./schemas/error_response/validationResponse.json#/definitions/validation_result"
                }
              }
            }
          }
        }
      }
    },
    "/{cats}/notices/{notice_id}": {
      "summary": "",
      "description": "",
      "parameters": [
        {
          "name": "cats",
          "in": "path",
          "description": "カフェ識別子",
          "required": true,
          "schema": {
            "type": "string",
            "enum": [
              "cats"
            ],
            "x-enum-varnames": [
              "cats"
            ]
          }
        },
        {
          "name": "notice_id",
          "in": "path",
          "description": "お知らせID",
          "required": true,
          "schema": {
            "type": "integer",
            "format": "int32"
          }
        }
      ],
      "get": {
        "tags": [
          "notice"
        ],
        "summary": "お知らせ詳細取得API",
        "description": "お知らせを1件取得する",
        "operationId": "noticeFind",
        "responses": {
          "200": {
            "description": "正常系",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./schemas/notice/response.json#/definitions/notice_find_response"
                }
              }
            }
          },
          "400": {
            "description": "バリデーションエラー",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./schemas/error_response/validationResponse.json#/definitions/validation_result"
                }
              }
            }
          },
          "404": {
            "description": "Idが見つからない"
          }
        }
      },
      "patch": {
        "tags": [
          "notice"
        ],
        "summary": "お知らせ更新API",
        "description": "お知らせを1件更新する",
        "operationId": "noticeUpdate",
        "requestBody": {
          "description": "リクエストボディ",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "./schemas/notice/request.json#/definitions/notice_update"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "正常系",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./schemas/success_response/patchResponse.json#/definitions/patch_object"
                }
              }
            }
          },
          "400": {
            "description": "バリデーションエラー",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./schemas/error_response/validationResponse.json#/definitions/validation_result"
                }
              }
            }
          },
          "404": {
            "description": "Idが見つからない"
          },
          "409": {
            "description": "排他失敗"
          }
        }
      },
      "delete": {
        "tags": [
          "notice"
        ],
        "summary": "お知らせ削除API",
        "description": "お知らせを1件論理削除する",
        "operationId": "noticeDelete",
        "parameters": [
          {
            "name": "version",
            "in": "query",
            "description": "バージョンフィールド",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "正常系"
          },
          "404": {
            "description": "Idが見つからない"
          },
          "409": {
            "description": "排他失敗"
          }
        }
      }
    },
...
    "/{cats}/provide_services": {
      "summary": "",
      "description": "",
      "parameters": [
        {
          "name": "cats",
          "in": "path",
          "description": "カフェ識別子",
          "required": true,
          "schema": {
            "type": "string",
            "enum": [
              "cats"
            ],
            "x-enum-varnames": [
              "cats"
            ]
          }
        }
      ],
      "get": {
        "tags": [
          "provide_service"
        ],
        "summary": "提供サービス一覧取得API",
        "description": "提供サービスを取得する",
        "operationId": "provideServiceSearch",
        "parameters": [
          {
            "name": "store_ids",
            "in": "query",
            "description": "店舗ID",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "integer",
                "format": "int32"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "正常系",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./schemas/provide_service/response.json#/definitions/provide_service_search_response"
                }
              }
            }
          },
          "400": {
            "description": "バリデーションエラー",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./schemas/error_response/validationResponse.json#/definitions/validation_result"
                }
              }
            }
          }
        }
      }
    },
    "/{cats}/stores/{store_id}/provide_service": {
      "summary": "",
      "description": "",
      "parameters": [
        {
          "name": "cats",
          "in": "path",
          "description": "カフェ識別子",
          "required": true,
          "schema": {
            "type": "string",
            "enum": [
              "cats"
            ],
            "x-enum-varnames": [
              "cats"
            ]
          }
        },
        {
          "name": "store_id",
          "in": "path",
          "description": "店舗ID",
          "required": true,
          "schema": {
            "type": "integer",
            "format": "int32"
          }
        }
      ],
      "post": {
        "tags": [
          "provide_service"
        ],
        "summary": "提供サービス登録API",
        "description": "提供サービスを1件登録する",
        "operationId": "provideServiceCreate",
        "requestBody": {
          "description": "リクエストボディ",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "./schemas/provide_service/request.json#/definitions/provide_service_create"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "正常系",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./schemas/success_response/postResponse.json#/definitions/post_object"
                }
              }
            }
          },
          "400": {
            "description": "バリデーションエラー",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./schemas/error_response/validationResponse.json#/definitions/validation_result"
                }
              }
            }
          }
        }
      }
    },
    "/{cats}/stores/{store_id}/provide_services/{provide_service_id}": {
      "summary": "",
      "description": "",
      "parameters": [
        {
          "name": "cats",
          "in": "path",
          "description": "カフェ識別子",
          "required": true,
          "schema": {
            "type": "string",
            "enum": [
              "cats"
            ],
            "x-enum-varnames": [
              "cats"
            ]
          }
        },
        {
          "name": "store_id",
          "in": "path",
          "description": "店舗ID",
          "required": true,
          "schema": {
            "type": "integer",
            "format": "int32"
          }
        },
        {
          "name": "provide_service_id",
          "in": "path",
          "description": "提供サービスID",
          "required": true,
          "schema": {
            "type": "integer",
            "format": "int32"
          }
        }
      ],
      "get": {
        "tags": [
          "provide_service"
        ],
        "summary": "提供サービス詳細取得API",
        "description": "提供サービス詳細を1件取得する",
        "operationId": "provideServiceFind",
        "responses": {
          "200": {
            "description": "正常系",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./schemas/provide_service/response.json#/definitions/provide_service_find_response"
                }
              }
            }
          },
          "400": {
            "description": "バリデーションエラー",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./schemas/error_response/validationResponse.json#/definitions/validation_result"
                }
              }
            }
          },
          "404": {
            "description": "Idが見つからない"
          }
        }
      },
      "patch": {
        "tags": [
          "provide_service"
        ],
        "summary": "提供サービス更新API",
        "description": "提供サービスを1件更新する",
        "operationId": "provideServiceUpdate",
        "requestBody": {
          "description": "リクエストボディ",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "./schemas/provide_service/request.json#/definitions/provide_service_update"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "正常系",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./schemas/success_response/patchResponse.json#/definitions/patch_object"
                }
              }
            }
          },
          "400": {
            "description": "バリデーションエラー",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./schemas/error_response/validationResponse.json#/definitions/validation_result"
                }
              }
            }
          },
          "404": {
            "description": "Idが見つからない"
          },
          "409": {
            "description": "排他失敗"
          }
        }
      },
      "delete": {
        "tags": [
          "provide_service"
        ],
        "summary": "提供サービス削除API",
        "description": "提供サービスを1件論理削除する",
        "operationId": "provideServiceDelete",
        "parameters": [
          {
            "name": "version",
            "in": "query",
            "description": "バージョンフィールド",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "正常系"
          },
          "404": {
            "description": "Idが見つからない"
          },
          "409": {
            "description": "排他失敗"
          }
        }
      }
    },
...
  }
}